requirements.txt
streamlit
openai
moviepy
python-dotenv